requirements.txt
datasets==2.20.0
pandas==2.2.2
mistralai==0.4.1
python-dotenv==1.0.1
scikit-learn==1.5.0
mistral-common==1.2.1
music21==9.1.0
matplotlib==3.9.0
PyAudio==0.2.14
Wave==0.0.2
pygame==2.6.0
gradio==4.37.2

